+++ romancal/tweakreg/tests/test_tweakreg.py
@@ -21,7 +21,7 @@ from gwcs.geometry import CartesianToSpherical, SphericalToCartesian
 from roman_datamodels import datamodels as rdm
 from roman_datamodels import maker_utils
 
-from romancal.datamodels import ModelContainer
+from romancal.datamodels import ModelLibrary
 from romancal.tweakreg import tweakreg_step as trs
 from romancal.tweakreg.astrometric_utils import get_catalog
 
@@ -475,20 +475,18 @@ def base_image():
 @pytest.mark.parametrize(
     "input, error_type",
     [
-        (list(), (TypeError,)),
-        ([""], (TypeError,)),
-        (["", ""], (TypeError,)),
-        ("", (TypeError,)),
-        ([1, 2, 3], (TypeError,)),
+        (list(), (Exception,)),
+        ([""], (Exception,)),
+        (["", ""], (Exception,)),
+        ("", (Exception,)),
+        ([1, 2, 3], (Exception,)),
     ],
 )
 def test_tweakreg_raises_error_on_invalid_input(input, error_type):
     # sourcery skip: list-literal
     """Test that TweakReg raises an error when an invalid input is provided."""
-    with pytest.raises(Exception) as exec_info:
+    with pytest.raises(error_type):
         trs.TweakRegStep.call(input)
-
-    assert type(exec_info.value) in error_type
 
 
 def test_tweakreg_raises_attributeerror_on_missing_tweakreg_catalog(base_image):
@@ -496,16 +494,14 @@ def test_tweakreg_raises_attributeerror_on_missing_tweakreg_catalog(base_image):
     Test that TweakReg raises an AttributeError if meta.tweakreg_catalog is missing.
     """
     img = base_image()
-    with pytest.raises(Exception) as exec_info:
+    with pytest.raises(AttributeError):
         trs.TweakRegStep.call([img])
 
-    assert type(exec_info.value) == AttributeError
 
-
-def test_tweakreg_returns_modelcontainer_on_roman_datamodel_as_input(
+def test_tweakreg_returns_modellibrary_on_roman_datamodel_as_input(
     tmp_path, base_image
 ):
-    """Test that TweakReg always returns a ModelContainer when processing an open Roman DataModel as input."""
+    """Test that TweakReg always returns a ModelLibrary when processing an open Roman DataModel as input."""
 
     img = base_image(shift_1=1000, shift_2=1000)
     add_tweakreg_catalog_attribute(tmp_path, img, catalog_filename="img_1")
@@ -513,29 +509,33 @@ def test_tweakreg_returns_modelcontainer_on_roman_datamodel_as_input(
     test_input = img
 
     res = trs.TweakRegStep.call(test_input)
-    assert res[0].meta.cal_step.tweakreg == "COMPLETE"
-    assert isinstance(res, ModelContainer)
+    assert isinstance(res, ModelLibrary)
+    with res:
+        model = res.borrow(0)
+        assert model.meta.cal_step.tweakreg == "COMPLETE"
+        res.shelve(model, 0, modify=False)
 
 
-def test_tweakreg_returns_modelcontainer_on_modelcontainer_as_input(
-    tmp_path, base_image
-):
-    """Test that TweakReg always returns a ModelContainer when processing a ModelContainer as input."""
+def test_tweakreg_returns_modellibrary_on_modellibrary_as_input(tmp_path, base_image):
+    """Test that TweakReg always returns a ModelLibrary when processing a ModelLibrary as input."""
 
     img = base_image(shift_1=1000, shift_2=1000)
     add_tweakreg_catalog_attribute(tmp_path, img, catalog_filename="img_1")
 
-    test_input = ModelContainer([img])
+    test_input = ModelLibrary([img])
 
     res = trs.TweakRegStep.call(test_input)
-    assert res[0].meta.cal_step.tweakreg == "COMPLETE"
-    assert isinstance(res, ModelContainer)
+    assert isinstance(res, ModelLibrary)
+    with res:
+        model = res.borrow(0)
+        assert model.meta.cal_step.tweakreg == "COMPLETE"
+        res.shelve(model, 0, modify=False)
 
 
-def test_tweakreg_returns_modelcontainer_on_association_file_as_input(
+def test_tweakreg_returns_modellibrary_on_association_file_as_input(
     tmp_path, base_image
 ):
-    """Test that TweakReg always returns a ModelContainer when processing an association file as input."""
+    """Test that TweakReg always returns a ModelLibrary when processing an association file as input."""
 
     img_1 = base_image(shift_1=1000, shift_2=1000)
     img_2 = base_image(shift_1=1000, shift_2=1000)
@@ -548,14 +548,17 @@ def test_tweakreg_returns_modelcontainer_on_association_file_as_input(
     test_input = asn_filepath
 
     res = trs.TweakRegStep.call(test_input)
-    assert all([x.meta.cal_step.tweakreg == "COMPLETE" for x in res])
-    assert isinstance(res, ModelContainer)
+    assert isinstance(res, ModelLibrary)
+    with res:
+        for i, model in enumerate(res):
+            assert model.meta.cal_step.tweakreg == "COMPLETE"
+            res.shelve(model, i, modify=False)
 
 
-def test_tweakreg_returns_modelcontainer_on_list_of_asdf_file_as_input(
+def test_tweakreg_returns_modellibrary_on_list_of_asdf_file_as_input(
     tmp_path, base_image
 ):
-    """Test that TweakReg always returns a ModelContainer when processing a list of ASDF files as input."""
+    """Test that TweakReg always returns a ModelLibrary when processing a list of ASDF files as input."""
 
     img_1 = base_image(shift_1=1000, shift_2=1000)
     img_2 = base_image(shift_1=1000, shift_2=1000)
@@ -571,14 +574,17 @@ def test_tweakreg_returns_modelcontainer_on_list_of_asdf_file_as_input(
     ]
 
     res = trs.TweakRegStep.call(test_input)
-    assert all([x.meta.cal_step.tweakreg == "COMPLETE" for x in res])
-    assert isinstance(res, ModelContainer)
+    assert isinstance(res, ModelLibrary)
+    with res:
+        for i, model in enumerate(res):
+            assert model.meta.cal_step.tweakreg == "COMPLETE"
+            res.shelve(model, i, modify=False)
 
 
-def test_tweakreg_returns_modelcontainer_on_list_of_roman_datamodels_as_input(
+def test_tweakreg_returns_modellibrary_on_list_of_roman_datamodels_as_input(
     tmp_path, base_image
 ):
-    """Test that TweakReg always returns a ModelContainer when processing a list of open Roman datamodels as input."""
+    """Test that TweakReg always returns a ModelLibrary when processing a list of open Roman datamodels as input."""
     img_1 = base_image(shift_1=1000, shift_2=1000)
     img_2 = base_image(shift_1=1000, shift_2=1000)
     add_tweakreg_catalog_attribute(tmp_path, img_1, catalog_filename="img_1")
@@ -589,8 +595,11 @@ def test_tweakreg_returns_modelcontainer_on_list_of_roman_datamodels_as_input(
     test_input = [img_1, img_2]
 
     res = trs.TweakRegStep.call(test_input)
-    assert all([x.meta.cal_step.tweakreg == "COMPLETE" for x in res])
-    assert isinstance(res, ModelContainer)
+    assert isinstance(res, ModelLibrary)
+    with res:
+        for i, model in enumerate(res):
+            assert model.meta.cal_step.tweakreg == "COMPLETE"
+            res.shelve(model, i, modify=False)
 
 
 def test_tweakreg_updates_cal_step(tmp_path, base_image):
@@ -599,8 +608,11 @@ def test_tweakreg_updates_cal_step(tmp_path, base_image):
     add_tweakreg_catalog_attribute(tmp_path, img)
     res = trs.TweakRegStep.call([img])
 
-    assert hasattr(res[0].meta.cal_step, "tweakreg")
-    assert res[0].meta.cal_step.tweakreg == "COMPLETE"
+    with res:
+        model = res.borrow(0)
+        assert hasattr(model.meta.cal_step, "tweakreg")
+        assert model.meta.cal_step.tweakreg == "COMPLETE"
+        res.shelve(model, 0, modify=False)
 
 
 def test_tweakreg_updates_group_id(tmp_path, base_image):
@@ -609,8 +621,10 @@ def test_tweakreg_updates_group_id(tmp_path, base_image):
     add_tweakreg_catalog_attribute(tmp_path, img)
     res = trs.TweakRegStep.call([img])
 
-    assert hasattr(res[0].meta, "group_id")
-    assert len(res[0].meta.group_id) > 0
+    with res:
+        model = res.borrow(0)
+        assert hasattr(model.meta, "group_id")
+        res.shelve(model, 0, modify=False)
 
 
 @pytest.mark.parametrize(
@@ -799,22 +813,24 @@ def test_tweakreg_combine_custom_catalogs_and_asn_file(tmp_path, base_image):
         catfile=catfile,
     )
 
-    assert type(res) == ModelContainer
+    assert type(res) == ModelLibrary
 
-    assert hasattr(res[0].meta, "asn")
+    with res:
+        for i, (model, target) in enumerate(zip(res, [img1, img2, img3])):
+            assert hasattr(model.meta, "asn")
 
-    assert all(
-        x.meta.asn["exptype"] == y["exptype"]
-        for x, y in zip(res, asn_content["products"][0]["members"])
-    )
+            assert (
+                model.meta["exptype"]
+                == asn_content["products"][0]["members"][i]["exptype"]
+            )
 
-    assert all(
-        x.meta.filename == y.meta.filename for x, y in zip(res, [img1, img2, img3])
-    )
+            assert model.meta.filename == target.meta.filename
 
-    assert all(type(x) == type(y) for x, y in zip(res, [img1, img2, img3]))
+            assert type(model) == type(target)
 
-    assert all((x.data == y.data).all() for x, y in zip(res, [img1, img2, img3]))
+            assert (model.data == target.data).all()
+
+            res.shelve(model, i, modify=False)
 
 
 @pytest.mark.parametrize(
@@ -978,28 +994,33 @@ def test_tweakreg_parses_asn_correctly(tmp_path, base_image):
         asn_content = json.load(f)
 
     res = trs.TweakRegStep.call(asn_filepath)
-    assert type(res) == ModelContainer
 
-    assert hasattr(res[0].meta, "asn")
-    assert (
-        res[0].meta.asn["exptype"]
-        == asn_content["products"][0]["members"][0]["exptype"]
-    )
-    assert (
-        res[1].meta.asn["exptype"]
-        == asn_content["products"][0]["members"][1]["exptype"]
-    )
-    assert res[0].meta.asn["pool_name"] == asn_content["asn_pool"]
-    assert res[1].meta.asn["pool_name"] == asn_content["asn_pool"]
+    assert type(res) == ModelLibrary
 
-    assert res[0].meta.filename == img_1.meta.filename
-    assert res[1].meta.filename == img_2.meta.filename
+    with res:
+        models = list(res)
+        assert hasattr(models[0].meta, "asn")
+        assert (
+            models[0].meta["exptype"]
+            == asn_content["products"][0]["members"][0]["exptype"]
+        )
+        assert (
+            models[1].meta["exptype"]
+            == asn_content["products"][0]["members"][1]["exptype"]
+        )
+        assert models[0].meta.asn["pool_name"] == asn_content["asn_pool"]
+        assert models[1].meta.asn["pool_name"] == asn_content["asn_pool"]
 
-    assert type(res[0]) == type(img_1)
-    assert type(res[1]) == type(img_2)
+        assert models[0].meta.filename == img_1.meta.filename
+        assert models[1].meta.filename == img_2.meta.filename
 
-    assert (res[0].data == img_1.data).all()
-    assert (res[1].data == img_2.data).all()
+        assert type(models[0]) == type(img_1)
+        assert type(models[1]) == type(img_2)
+
+        assert (models[0].data == img_1.data).all()
+        assert (models[1].data == img_2.data).all()
+
+        [res.shelve(m, i, modify=False) for i, m in enumerate(models)]
 
 
 def test_tweakreg_raises_error_on_connection_error_to_the_vo_service(
@@ -1016,9 +1037,12 @@ def test_tweakreg_raises_error_on_connection_error_to_the_vo_service(
     monkeypatch.setattr("requests.get", MockConnectionError)
     res = trs.TweakRegStep.call([img])
 
-    assert type(res) == ModelContainer
+    assert type(res) == ModelLibrary
     assert len(res) == 1
-    assert res[0].meta.cal_step.tweakreg.lower() == "skipped"
+    with res:
+        model = res.borrow(0)
+        assert model.meta.cal_step.tweakreg.lower() == "skipped"
+        res.shelve(model, 0, modify=False)
 
 
 def test_fit_results_in_meta(tmp_path, base_image):
@@ -1031,11 +1055,12 @@ def test_fit_results_in_meta(tmp_path, base_image):
 
     res = trs.TweakRegStep.call([img])
 
-    assert type(res) == ModelContainer
-    assert [
-        hasattr(x.meta, "wcs_fit_results") and len(x.meta.wcs_fit_results) > 0
-        for x in res
-    ]
+    assert type(res) == ModelLibrary
+    with res:
+        for i, model in enumerate(res):
+            assert hasattr(model.meta, "wcs_fit_results")
+            assert len(model.meta.wcs_fit_results) > 0
+            res.shelve(model, i, modify=False)
 
 
 def test_tweakreg_handles_multiple_groups(tmp_path, base_image):
@@ -1056,14 +1081,14 @@ def test_tweakreg_handles_multiple_groups(tmp_path, base_image):
 
     res = trs.TweakRegStep.call([img1, img2])
 
-    assert len(res.models_grouped) == 2
-    all(
-        (
-            r.meta.group_id.split("-")[1],
-            i.meta.observation.program.split("-")[1],
-        )
-        for r, i in zip(res, [img1, img2])
-    )
+    assert len(res.group_names) == 2
+    with res:
+        for r, i in zip(res, [img1, img2]):
+            assert (
+                r.meta.group_id.split("-")[1]
+                == i.meta.observation.program.split("-")[1]
+            )
+            res.shelve(r, modify=False)
 
 
 @pytest.mark.parametrize(
@@ -1086,19 +1111,21 @@ def test_imodel2wcsim_valid_column_names(tmp_path, base_image, column_names):
             format=catalog_format,
         )
         x.meta.tweakreg_catalog.rename_columns(("x", "y"), column_names)
+    xname, yname = column_names
 
-    images = ModelContainer([img_1, img_2])
-    grp_img = list(images.models_grouped)
-    g = grp_img[0]
+    images = ModelLibrary([img_1, img_2])
 
     step = trs.TweakRegStep()
-    imcats = list(map(step._imodel2wcsim, g))
-
-    assert all(x.meta["image_model"]() == y for x, y in zip(imcats, [img_1, img_2]))
-    assert np.all(
-        x.meta["catalog"] == y.meta.tweakreg_catalog
-        for x, y in zip(imcats, [img_1, img_2])
-    )
+    with images:
+        for i, (m, target) in enumerate(zip(images, [img_1, img_2])):
+            imcat = step._imodel2wcsim(m)
+            assert (
+                imcat.meta["catalog"]["x"] == target.meta.tweakreg_catalog[xname]
+            ).all()
+            assert (
+                imcat.meta["catalog"]["y"] == target.meta.tweakreg_catalog[yname]
+            ).all()
+            images.shelve(m, i, modify=False)
 
 
 @pytest.mark.parametrize(
@@ -1125,15 +1152,14 @@ def test_imodel2wcsim_error_invalid_column_names(tmp_path, base_image, column_na
         )
         x.meta.tweakreg_catalog.rename_columns(("x", "y"), column_names)
 
-    images = ModelContainer([img_1, img_2])
-    grp_img = list(images.models_grouped)
-    g = grp_img[0]
+    images = ModelLibrary([img_1, img_2])
 
     step = trs.TweakRegStep()
-    with pytest.raises(Exception) as exec_info:
-        list(map(step._imodel2wcsim, g))
-
-    assert type(exec_info.value) == ValueError
+    with pytest.raises(ValueError):
+        with images:
+            for i, model in enumerate(images):
+                images.shelve(model, i, modify=False)
+                step._imodel2wcsim(model)
 
 
 def test_imodel2wcsim_error_invalid_catalog(tmp_path, base_image):
@@ -1145,15 +1171,14 @@ def test_imodel2wcsim_error_invalid_catalog(tmp_path, base_image):
     # set meta.tweakreg_catalog (this is automatically added by TweakRegStep)
     img_1.meta["tweakreg_catalog"] = "nonsense"
 
-    images = ModelContainer([img_1])
-    grp_img = list(images.models_grouped)
-    g = grp_img[0]
+    images = ModelLibrary([img_1])
 
     step = trs.TweakRegStep()
-    with pytest.raises(Exception) as exec_info:
-        list(map(step._imodel2wcsim, g))
-
-    assert type(exec_info.value) == AttributeError
+    with pytest.raises(AttributeError):
+        with images:
+            for i, model in enumerate(images):
+                images.shelve(model, i, modify=False)
+                step._imodel2wcsim(model)
 
 
 def test_parse_catfile_valid_catalog(tmp_path, base_image):